Extract number formatter and simplify formats
Move formatNumber into a shared lib (app/src/lib/format-number.ts) and remove the duplicate implementation from formatting-preferences. Simplify numberFormats to be a ReadonlyArray<INumberFormat> (remove precomputed preview/example), update numberFormatFromKey and lookup logic to account for the new shape, and update the preferences UI to import and use formatNumber for previews and to pass formats directly on change. This refactors formatting logic for reuse, reduces duplication, and keeps the UI responsible for rendering preview strings.
